feat(BackgroundProgress): Allow passing busyObject
This change allows the `BackgroundProgress` widget to display a loading state based on a specific `busyObject` in addition to the global `isBusy` state from the view model.

If `busyObject` is provided, the widget will show the progress indicator when `viewModel.busy(busyObject)` is true. Otherwise, it will fallback to `viewModel.isBusy`.